Split tests in different files
